requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.6
django-cors-headers==4.7.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
drf-yasg==1.21.8
filetype==1.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.68.0
grpcio==1.70.0
grpcio-status==1.70.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
legacy-cgi==2.6.2
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==24.2
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.11.0a2
pydantic_core==2.29.0
pyparsing==3.2.1
pypdf==5.3.0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
setuptools==75.8.0
simplejson==3.19.3
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
websockets==14.2
